+++ waspc/e2e-tests/test-outputs/snapshots/wasp-build-golden/wasp-app/.wasp/out/sdk/wasp/client/app/router/router.tsx
@@ -10,7 +10,7 @@ export function getRouter({
   routesMapping,
   rootElement,
 }: {
-  routesMapping: Record<string, React.ComponentType>,
+  routesMapping: Record<string, { lazy: () => Promise<{ Component: React.ComponentType }> } | { Component: React.ComponentType }>,
   rootElement: React.ReactNode,
 }) {
   const waspDefinedRoutes = [
@@ -18,7 +18,7 @@ export function getRouter({
   const userDefinedRoutes = Object.entries(routes).map(([routeKey, route]) => {
     return {
       path: route.to,
-      Component: routesMapping[routeKey],
+      ...routesMapping[routeKey],
     }
   })