+++ waspc/e2e-tests/test-outputs/snapshots/wasp-compile-golden/wasp-app/.wasp/out/sdk/wasp/client/app/components/WaspApp.tsx
@@ -7,7 +7,7 @@ import { queryClientInitialized } from '../../operations/index'
 
 export type WaspAppProps = {
   rootElement?: React.ReactNode;
-  routesMapping: Record<string, React.ComponentType>;
+  routesMapping: Record<string, { lazy: () => Promise<{ Component: React.ComponentType }> } | { Component: React.ComponentType }>;
 }
 
 export function WaspApp({ rootElement, routesMapping }: Required<WaspAppProps>) {